+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessJSFilesTask.java
@@ -35,6 +35,7 @@ import com.google.common.collect.Lists;
 import com.google.javascript.jscomp.Compiler;
 import com.google.javascript.jscomp.CompilerOptions;
 import com.google.javascript.jscomp.SourceFile;
+import com.google.javascript.rhino.head.EvaluatorException;
 import com.samaxes.maven.minify.common.ClosureConfig;
 import com.samaxes.maven.minify.common.JavaScriptErrorReporter;
 import com.samaxes.maven.minify.common.YuiConfig;
@@ -113,6 +114,11 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
 
                     Compiler compiler = new Compiler();
                     compiler.compile(externs, Lists.newArrayList(input), options);
+
+                    if (compiler.hasErrors()) {
+                        throw new EvaluatorException(compiler.getErrors()[0].description);
+                    }
+
                     writer.append(compiler.toSource());
                     break;
                 case YUI: